+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # CHANGELOG.md
 
+## Unreleased
+
+- Add `-p,--prettier` option to the cli to format the generated code with prettier [#51](https://github.com/Embraser01/typoas/pull/51)
+
 ## 3.1.7 - 2024-02-23
+
 ## 3.1.6 - 2024-02-23
 
 > Only the runtime was published, the generator and cli are still at 3.1.5